+++ app/src/main/java/edu/gatech/cs2340/myapplication/controllers/fragments/RegisterFragment.java
@@ -1,8 +1,7 @@
-package edu.gatech.cs2340.myapplication;
+package edu.gatech.cs2340.myapplication.controllers.fragments;
 
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,7 +15,9 @@ import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textfield.TextInputLayout;
 
 import androidx.fragment.app.Fragment;
-import androidx.navigation.Navigation;
+import edu.gatech.cs2340.myapplication.R;
+import edu.gatech.cs2340.myapplication.models.The_Cloud;
+import edu.gatech.cs2340.myapplication.models.User;
 
 /**
  * A simple {@link Fragment} subclass.
