+++ app/src/main/res/layout/fragment_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainFragment">
+    tools:context=".controllers.fragments.MainFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
