+++ app/src/main/res/layout/fragment_view_location.xml
@@ -8,7 +8,7 @@
     android:paddingStart="4dp"
     android:paddingEnd="4dp"
     android:background="@color/colorPrimary"
-    tools:context=".ViewLocationFragment">
+    tools:context=".controllers.fragments.ViewLocationFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
